requirements.txt
streamlit
PyMuPDF
google-generativeai
python-docx
python-dotenv
beautifulsoup4